Add quick way of adding applications
