+++ app/views/applications/interview-needs.html
@@ -25,7 +25,7 @@
             id: "interview-needs-detail",
             value: data.applications[id].interviewNeedsDetail,
             label: {
-              text: "Tell us why you need to be flexible",
+              text: "Tell us what flexibility you need",
               classes: "govuk-label--s"
             },
             maxlength: 150,
@@ -39,7 +39,7 @@
           fieldset: {
             legend: {
               classes: "govuk-fieldset__legend--m",
-              text: "Do you need to be flexible with your interview time and date?"
+              text: "Do you need flexibility with your interview time and date?"
             }
           },
           items: [{
