+++ app/views/layouts/main.html
@@ -54,6 +54,10 @@
           text: "Complete your details"
         },
         {
+          href: "/admin/add-application",
+          text: "Add application"
+        },
+        {
           href: "/admin/receive-offer",
           text: "Receive an offer"
         },
